hireon-backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.38
asyncpg>=0.29.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Validation
pydantic
pydantic-settings==2.3.0
email-validator==2.1.1

# AI - Groq & Gemini (free)
google-generativeai==0.7.2
groq==1.1.0
mistralai==1.1.0
numpy

# File parsing
pdfplumber>=0.11.0
PyPDF2==3.0.1
python-docx==1.1.2

# PDF generation
xhtml2pdf==0.2.16

# HTTP client
httpx==0.27.0

# Utils
python-dateutil==2.8.2
pillow
aiofiles==23.2.1

# Google APIs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
